cloud_run/pyproject.toml
[project]
name = "deploy_vertex"
version = "0.0.1"
description = "Se deplega docker en vertex"
requires-python = ">=3.10,<3.13"
dependencies = [
   "fastapi[all]>=0.89.1",
   "uvicorn[standard]>=0.20.0",
   "loguru>=0.7.0",
   "pillow>=9.0.0",
   "ultralytics>=8.3.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"